+++ HighPingKicker/Service.cs
@@ -38,36 +38,43 @@ namespace HighPingKicker {
             var key = ClientToId(clientInfo);
             log.Debug("Ping Check: {ClientToId(clientInfo)}({clientInfo.playerName}): {ping}ms"); // TODO: remove
             if (ping <= Config.MaxPingAllowed) {
-                Decrement(PingCounters, key);
+                var reducedFailureBudget = Decrement(PingCounters, key);
+                log.Info($"Successful ping for {ClientToId(clientInfo)}({clientInfo.playerName}): {ping}ms <= {Config.MaxPingAllowed}ms. Ping failure budget recovering: {reducedFailureBudget}/{Config.FailureThresholdBeforeKick}.");
                 return;
             }
 
-            if (IncrementThenGet(PingCounters, key) > Config.FailureThresholdBeforeKick) {
+            var failureBudget = IncrementThenGet(PingCounters, key);
+            log.Info($"Ping Failure for {ClientToId(clientInfo)}({clientInfo.playerName}): {ping}ms > {Config.MaxPingAllowed}ms. Ping failure budget impacted: {failureBudget}/{Config.FailureThresholdBeforeKick}.");
+            if (failureBudget > Config.FailureThresholdBeforeKick) {
                 if (IncrementThenGet(KickCounters, key) > Config.AllowedKicksBeforeBan) {
                     var banExpiration = DateTime.Now.AddHours(Config.HoursBannedAfterKickWarnings);
-                    var reason = $"kicked multiple times for excessive latency above {Config.MaxPingAllowed}ms but the problem continues to persist";
-                    Ban(clientInfo, banExpiration, reason);
-                    log.Warn($"{clientInfo.playerName}/{ClientToId(clientInfo)} was banned: {reason}");
+                    KickAndBan(clientInfo, banExpiration);
+                    log.Warn($"{clientInfo.playerName}/{ClientToId(clientInfo)} banned. Ping: {ping}");
+                    BroadcastMessage(clientInfo, $"{clientInfo.playerName} was automatically banned for {Config.HoursBannedAfterKickWarnings}hrs after being auto-kicked multiple times for excessive latency above {Config.MaxPingAllowed}ms.");
+                    PingCounters.Remove(key);
+                    KickCounters.Remove(key);
 
                     // Don't ban family accounts... for now
                     // BanFamilyAccount(clientInfo, banExpiration, reason);
                 } else {
-                    var reason = $"latency exceeded the limit of {Config.MaxPingAllowed}ms multiple times, but can reconnect after checking router/network";
+                    var reason = $"Your connection to us exceeded the latency limit of {Config.MaxPingAllowed}ms multiple times, so you were automatically kicked from the server. Consider checking your router/computer/network to ensure your connection to us is functioning properly before attempting to reconnect. You can reconnect when ready, but please be aware that being kicked multiple times for this issue may result in a ban.";
                     Kick(clientInfo, reason);
-                    log.Warn($"{clientInfo.playerName}/{ClientToId(clientInfo)} was kicked: {reason}");
+                    log.Warn($"{clientInfo.playerName}/{ClientToId(clientInfo)} kicked. Ping: {ping}");
+                    BroadcastMessage(clientInfo, $"{clientInfo.playerName} was automatically kicked exceeded the latency limit of {Config.MaxPingAllowed}ms multiple times.");
+                    PingCounters.Remove(key);
                 }
             }
         }
 
         private void Kick(ClientInfo clientInfo, string reason, GameUtils.EKickReason kickReason = GameUtils.EKickReason.ModDecision) {
+            log.Info($"Kicked {ClientToId(clientInfo)}({clientInfo.playerName})");
             GameUtils.KickPlayerForClientInfo(clientInfo, new GameUtils.KickPlayerData(kickReason, 0, default, reason));
-            BroadcastMessage(clientInfo, $"{clientInfo.playerName} was kicked: {reason}");
         }
 
-        private void Ban(ClientInfo clientInfo, DateTime banExpiration, string reason) {
-            Kick(clientInfo, reason, GameUtils.EKickReason.Banned);
-            BroadcastMessage(clientInfo, $"{clientInfo.playerName} was banned for {Config.HoursBannedAfterKickWarnings}hrs: {reason}");
-            GameManager.Instance.adminTools.AddBan(clientInfo.playerName, clientInfo.PlatformId, banExpiration, "Excessive Network Latency; Please check your router and network before connecting again or consider finding another server if this has happened multiple times.", true);
+        private void KickAndBan(ClientInfo clientInfo, DateTime banExpiration) {
+            log.Info($"Banned {ClientToId(clientInfo)}({clientInfo.playerName})");
+            Kick(clientInfo, "You were banned for excessive Network Latency and can try connecting again after the indicated time shown here. If you do not believe there is an ongoing problem with your network, you may simply be located too far away from this server. In that case, we would recommend trying a different server.", GameUtils.EKickReason.Banned);
+            GameManager.Instance.adminTools.AddBan(clientInfo.playerName, clientInfo.PlatformId, banExpiration, $"Banned for {Config.HoursBannedAfterKickWarnings}hrs after auto-kicked multiple times for latency above {Config.MaxPingAllowed}ms.", true);
         }
 
         private void BanFamilyAccount(ClientInfo clientInfo, DateTime banExpiration, string reason) {
@@ -90,13 +97,16 @@ namespace HighPingKicker {
             return counters[key];
         }
 
-        private void Decrement(Dictionary<string, int> counters, string key) {
+        private int Decrement(Dictionary<string, int> counters, string key) {
             if (counters.ContainsKey(key)) {
                 counters[key]--;
                 if (counters[key] == 0) {
                     counters.Remove(key);
+                    return 0;
                 }
+                return counters[key];
             }
+            return 0;
         }
 
         private string ClientToId(ClientInfo clientInfo) {