+++ CHANGELOG.md
@@ -7,11 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [TODO]
 
-- add command to show current ping and kick counters
 - add option for system to whisper to users about latency
+- redo text so it fits in message box
 
 ## [1.0.0] - 2022-05-13
 
+- add command to show current ping and kick counters
 - add logs entries any time a counter increases
 - flush ping counter when players is kicked
 - flush ping & kick counter when player is banned