+++ CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2022-05-13
+## [1.0.1] - 2022-05-15
+
+- fix logic issue when good ping is detected
+
+## [1.0.0] - 2022-05-14
 
 - add command to show current ping and kick counters
 - add logs entries any time a counter increases